feat-role based auth

+++ src/main/java/com/crud/CrudApp/Journal/JournalService.java
@@ -3,6 +3,8 @@ package com.crud.CrudApp.Journal;
 import com.crud.CrudApp.User.User;
 import com.crud.CrudApp.User.UserService;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,6 +15,7 @@ import java.util.Optional;
 
 @Service
 public class JournalService {
+    private static final Logger log = LoggerFactory.getLogger(JournalService.class);
     @Autowired
     private JournalRepository journalRepository;
     @Autowired
@@ -37,16 +40,22 @@ public class JournalService {
         return journalRepository.findById(String.valueOf(id));
     }
 
-    public void deleteById(ObjectId id, String userName) {
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
         try {
             User user = userService.getByUsername(userName);
-            user.getJournalEntries().removeIf(journal -> journal.getId().equals(id));
-            userService.saveUser(user);
-            journalRepository.deleteById(String.valueOf(id));
+            removed = user.getJournalEntries().removeIf(journal -> journal.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalRepository.deleteById(String.valueOf(id));
+            }
         }
         catch (Exception e){
+            log.error(e.getMessage());
             throw new RuntimeException("Error while deleting journal./n", e);
         }
+        return removed;
     }
 
     public void update(Journal journal) {
